+++ src/list-item.tsx
@@ -3,10 +3,6 @@ import { AppEntry } from "./types";
 import { restartApp } from "./utils";
 
 export function ListItem({ app }: { app: AppEntry }) {
-  function onRestart() {
-    restartApp(app.name);
-  }
-
   return (
     <List.Item
       title={app.name}
@@ -15,7 +11,7 @@ export function ListItem({ app }: { app: AppEntry }) {
       icon={{ fileIcon: app.iconPath }}
       actions={
         <ActionPanel>
-          <Action title="Restart App" onAction={onRestart} />
+          <Action title="Restart App" onAction={() => restartApp(app.name)} />
         </ActionPanel>
       }
     />